+++ releases/changes.md
@@ -1,2 +1,4 @@
 * Markdown documentation improvements
-* Splashscreen for fu-virfilter fixed
+* Splashscreen for *fu-virfilter* fixed
+* Argument parser for _fu-cov_ improved
+* Now `seqfu --version` and `seqfu version` will print the version number and exit